web/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psutil>=6.0.0
jinja2>=3.1.0
python-multipart>=0.0.18
pydantic-settings>=2.7.0
websockets>=14.0
